+++ gleam_stdlib/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Labels have been added to functions throughout the stdlib.
 - `map.fetch` has been renamed to `map.get` and `map.put` to `map.insert`.
+- `list.find` has been renamed `list.find_map` and a new `list.find` has been
+  introduced.
 - The `pair` module gains the `map_first`, and `map_second` functions.
 - The `string` module gains the `compare` function.
 - The `float` module gains the `max`,  and `min` functions.